(#16) refactor: nav 서버 컴포넌트 변경
