+++ src/components/layout/layoutNav.js
@@ -1,26 +1,16 @@
-"use client";
-import React, { useState } from "react";
+import React from "react";
 import s from "./layoutNav.module.css";
 import Link from "next/link";
-import { signIn, signOut, useSession } from "next-auth/react";
-import Modal from "../modal/modal";
+import SignInBtn from "./signinbtn";
+import SignOutBtn from "./signoutbtn";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
-export default function LayoutNav() {
-  const [loginModalOpen, setLoginModalOpen] = useState(false);
-  const [logoutModalOpen, setLogoutModalOpen] = useState(false);
-  const { data: session } = useSession();
-
-  const handleLogout = async () => {
-    await signOut();
-    window.location.href = "/";
-  };
-
+export default async function LayoutNav() {
+  const session = await getServerSession(authOptions);
   return (
     <nav className={s.nav}>
       <ul>
-        <li>
-          <Link href="/">Go Home</Link>
-        </li>
         {session && (
           <li>
             <Link href="/timecheck">Time check</Link>
@@ -28,61 +18,14 @@ export default function LayoutNav() {
         )}
         {session ? (
           <li>
-            <button
-              type="button"
-              onClick={() => setLogoutModalOpen(true)}
-              className={s.button}
-            >
-              Sign out
-            </button>
+            <SignOutBtn />
           </li>
         ) : (
           <li>
-            <button
-              type="button"
-              onClick={() => {
-                setLoginModalOpen(true);
-              }}
-              className={s.button}
-            >
-              Sign In
-            </button>
+            <SignInBtn />
           </li>
         )}
       </ul>
-      {loginModalOpen && (
-        <Modal closeModal={() => setLoginModalOpen(false)}>
-          <strong className={s.loginStrong}>Login</strong>
-          <button
-            type="button"
-            onClick={() => signIn("kakao")}
-            className={s.loginBtn}
-          >
-            Kakao 로그인
-          </button>
-        </Modal>
-      )}
-      {logoutModalOpen && (
-        <Modal closeModal={() => setLogoutModalOpen(false)}>
-          <p>로그아웃 하시겠습니까?</p>
-          <div className={s.logoutBtnDiv}>
-            <button
-              type="button"
-              onClick={() => setLogoutModalOpen(false)}
-              className={s.logoutBtn}
-            >
-              아니요
-            </button>
-            <button
-              type="button"
-              onClick={handleLogout}
-              className={s.logoutBtn}
-            >
-              네
-            </button>
-          </div>
-        </Modal>
-      )}
     </nav>
   );
 }
